+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - A method to get a balance
 - Metric collecting
 - Chain ID field to the Subscription type in .did interfaces
+- Executions counter to Subscriptions
 
 ### Fixed
 - Refresh README.md